requirements.txt
pandas
kafka-python
xgboost
graphviz
math
elasticsearch
